requirements.txt
fastapi
uvicorn
jinja2
python-multipart
passlib[bcrypt]
pydantic
pymongo
htmx
python-dotenv
bcrypt
